rename some functions; add docstrings; small refactoring
