+++ wikitools/errors.py
@@ -9,6 +9,10 @@ class LinkError:
 
 
 class LinkNotFound(LinkError, collections.namedtuple('LinkNotFound', 'location')):
+    """
+    An error indicating missing link: a text or binary file does not exist, and there is no such redirect.
+    """
+
     location: str
 
     def __repr__(self):
@@ -19,6 +23,10 @@ class BrokenRedirect(
     LinkError,
     collections.namedtuple('BrokenRedirect', 'location redirect_lineno redirect_destination')
 ):
+    """
+    An error indicating broken redirect: an article from the redirect.yaml file does not exist.
+    """
+
     location: str
     redirect_lineno: int
     redirect_destination: str
@@ -30,6 +38,11 @@ class BrokenRedirect(
 
 
 class MissingReference(LinkError, collections.namedtuple('MissingReference', 'location')):
+    """
+    An error indicating that a reference-style link is missing its counterpart:
+    [link][link_ref] exists, but [link_ref]: /wiki/Path/To/Article does not.
+    """
+
     location: str
 
     def __repr__(self):
